requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.3
pdfminer.six==20221105
requests==2.31.0
edge-tts==6.1.8
asyncio==3.4.3
aiofiles==23.2.1
groq==0.4.1
openai==1.6.1          # UPDATED to meet langchain-openai requirement
langchain==0.0.340
langchain-openai==0.0.2
langchain-groq==0.0.1
python-multipart==0.0.6
